+++ src/features/consultation/components/ConsultationPage.tsx
@@ -4,6 +4,7 @@ import { ConsultationHero } from "./ConsultationHero";
 import { ConsultationContent } from "./ConsultationContent";
 import { ConsultationProcess } from "./ConsultationProcess";
 import { ConsultationCompetency } from "./ConsultationCompetency";
+import { ProcessTimeline } from "./ProcessTimeline";
 import { ConsultationCTA } from "@/features/layout/components/ConsultationCTA";
 
 export const ConsultationPage = () => {
@@ -11,6 +12,7 @@ export const ConsultationPage = () => {
         <main className="flex flex-col w-full">
             <ConsultationHero />
             <ConsultationContent />
+            <ProcessTimeline/>
             <ConsultationProcess />
             <ConsultationCompetency />
             <ConsultationCTA />